requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.3.1
    # via
    #   tensorboard
    #   tensorflow-macos
anyio==4.9.0
    # via starlette
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
astunparse==1.6.3
    # via tensorflow-macos
blis==1.2.1
    # via thinc
boto3==1.39.3
    # via -r requirements.in
botocore==1.39.3
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.7.9
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   typer
    #   uvicorn
cloudpathlib==0.21.1
    # via weasel
comm==0.2.2
    # via ipykernel
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
exceptiongroup==1.2.2
    # via
    #   anyio
    #   ipython
executing==2.2.0
    # via stack-data
fastapi==0.78.0
    # via -r requirements.in
flatbuffers==25.2.10
    # via tensorflow-macos
fonttools==4.58.5
    # via matplotlib
from-root==1.3.0
    # via -r requirements.in
gast==0.4.0
    # via tensorflow-macos
google-auth==2.40.3
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==1.0.0
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow-macos
grpcio==1.73.1
    # via
    #   tensorboard
    #   tensorflow-macos
h11==0.16.0
    # via uvicorn
h5py==3.14.0
    # via tensorflow-macos
idna==3.10
    # via
    #   anyio
    #   requests
ipykernel==6.29.5
    # via -r requirements.in
ipython==8.23.0
    # via ipykernel
jedi==0.19.2
    # via ipython
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   spacy
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.1
    # via
    #   nltk
    #   scikit-learn
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
keras==2.13.1
    # via tensorflow-macos
kiwisolver==1.4.8
    # via matplotlib
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
libclang==18.1.1
    # via tensorflow-macos
marisa-trie==1.2.1
    # via language-data
markdown==3.8.2
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.3
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
mypy-boto3-s3==1.39.2
    # via -r requirements.in
nest-asyncio==1.6.0
    # via ipykernel
nltk==3.9.1
    # via -r requirements.in
numpy==1.24.3
    # via
    #   -r requirements.in
    #   blis
    #   contourpy
    #   h5py
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   spacy
    #   tensorboard
    #   tensorflow-macos
    #   thinc
oauthlib==3.3.1
    # via requests-oauthlib
opt-einsum==3.4.0
    # via tensorflow-macos
packaging==25.0
    # via
    #   ipykernel
    #   matplotlib
    #   spacy
    #   tensorflow-macos
    #   thinc
    #   weasel
pandas==2.3.1
    # via seaborn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via matplotlib
platformdirs==4.3.8
    # via jupyter-core
preshed==3.0.10
    # via
    #   spacy
    #   thinc
prompt-toolkit==3.0.51
    # via ipython
protobuf==4.25.8
    # via
    #   tensorboard
    #   tensorflow-macos
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==1.10.22
    # via
    #   confection
    #   fastapi
    #   spacy
    #   thinc
    #   weasel
pygments==2.19.2
    # via
    #   ipython
    #   rich
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
regex==2024.11.6
    # via
    #   -r requirements.in
    #   nltk
requests==2.32.4
    # via
    #   requests-oauthlib
    #   spacy
    #   tensorboard
    #   weasel
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.0.0
    # via typer
rsa==4.9.1
    # via google-auth
s3transfer==0.13.0
    # via boto3
scikit-learn==1.7.0
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
seaborn==0.13.2
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow-macos
    #   tensorflow-metal
smart-open==7.3.0.post1
    # via weasel
sniffio==1.3.1
    # via anyio
spacy==3.8.7
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
stack-data==0.6.3
    # via ipython
starlette==0.19.1
    # via fastapi
tensorboard==2.13.0
    # via tensorflow-macos
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow-estimator==2.13.0
    # via tensorflow-macos
tensorflow-macos==2.13.0
    # via -r requirements.in
tensorflow-metal==1.2.0
    # via -r requirements.in
termcolor==3.1.0
    # via tensorflow-macos
thinc==8.3.4
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via
    #   nltk
    #   spacy
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typer==0.16.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.5.0
    # via
    #   anyio
    #   cloudpathlib
    #   ipython
    #   mypy-boto3-s3
    #   pydantic
    #   rich
    #   tensorflow-macos
    #   typer
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   requests
uvicorn==0.18.3
    # via -r requirements.in
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
wcwidth==0.2.13
    # via prompt-toolkit
weasel==0.4.1
    # via spacy
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.1
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow-metal
wrapt==1.17.2
    # via
    #   smart-open
    #   tensorflow-macos

# The following packages are considered to be unsafe in a requirements file:
# setuptools
